+++ Sources/Network/GrpcConnection/ArbitraryGrpcConnection.swift
@@ -50,11 +50,11 @@ class ArbitraryGrpcConnection {
 extension ArbitraryGrpcConnection {
     private struct Inner {
         let url: MobileCoinUrlProtocol
-        private let session: GrpcConnectionSession
+        private let session: ConnectionSession
 
         init(url: MobileCoinUrlProtocol) {
             self.url = url
-            self.session = GrpcConnectionSession(url: url)
+            self.session = ConnectionSession(url: url)
         }
 
         func requestCallOptions() -> CallOptions {
